Cleaned up code for cubeItems and 3ditems. added various items.

+++ mod/threeDitems/helper/ArmorHelper.java
@@ -40,21 +40,21 @@ public class ArmorHelper {
 				model.bipedRightLeg.showModel = for2 || for3;
 				model.bipedLeftLeg.showModel = for2 || for3;
 
-				
+
 				float f1 = 1.0F;
 				GL11.glColor3f(f1, f1, f1);
 
 				if (itemarmor.getArmorMaterial() == EnumArmorMaterial.CLOTH)
 				{
 					int j = itemarmor.getColor(itemstack);
-					float f2 = (float)(j >> 16 & 255) / 255.0F;
-					float f3 = (float)(j >> 8 & 255) / 255.0F;
-					float f4 = (float)(j & 255) / 255.0F;
-//					GL11.glColor3f(0f,0f,1f);
+					float f2 = ((j >> 16) & 255) / 255.0F;
+					float f3 = ((j >> 8) & 255) / 255.0F;
+					float f4 = (j & 255) / 255.0F;
+					//					GL11.glColor3f(0f,0f,1f);
 
 					GL11.glColor3f(f1 * f2, f1 * f3, f1 * f4);
 				}
-//				
+				//
 			}
 		}
 	}
